+++ src/main/java/kg/peaksoft/ebookb4/api/AuthController.java
@@ -1,6 +1,7 @@
 package kg.peaksoft.ebookb4.api;
 
 import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.tags.Tag;
 import kg.peaksoft.ebookb4.dto.request.LoginRequest;
 import kg.peaksoft.ebookb4.dto.request.SignupRequestClient;
 import kg.peaksoft.ebookb4.dto.request.SignupRequestVendor;
@@ -26,6 +27,7 @@ import java.util.stream.Collectors;
 @RestController
 @RequestMapping("/api/auth")
 @AllArgsConstructor
+@Tag(name = "AuthController",description = "jwt token,")
 public class AuthController {
 
     AuthenticationManager authenticationManager;
@@ -53,15 +55,7 @@ public class AuthController {
                 userDetails.getEmail(),
                 roles));
     }
-    @Operation(summary = "singup",description = "registered cliented")
-    @PostMapping("/signup/client")
-    public ResponseEntity<?> registerClient(@Valid @RequestBody SignupRequestClient client) {
-        return clientService.register(client, 1L);
-    }
-    @Operation(summary = "signup",description = "registreted vendor")
-    @PostMapping("/signup/vendor")
-    public ResponseEntity<?> registerVendor(@Valid @RequestBody SignupRequestVendor vendor) {
-        return vendorService.register(vendor, 2L);
-    }
+
+
 
 }
