+++ src/main/java/kg/peaksoft/ebookb4/config/WebSecurityConfig.java
@@ -75,6 +75,7 @@ public class WebSecurityConfig extends WebSecurityConfigurerAdapter {
       .sessionManagement().sessionCreationPolicy(SessionCreationPolicy.STATELESS).and()
       .authorizeRequests().antMatchers("/api/auth/**", "/").permitAll()
       .antMatchers("/api/test/**").permitAll()
+            .antMatchers(AUTH_WHITELIST).permitAll()
       .anyRequest().authenticated();
 
     http.addFilterBefore(authenticationJwtTokenFilter(), UsernamePasswordAuthenticationFilter.class);
