+++ src/main/resources/application-dev.properties
@@ -1,7 +1,7 @@
 #Datasource conf
 spring.datasource.url=jdbc:postgresql://localhost:5432/ebook
 spring.datasource.username=postgres
-spring.datasource.password=postgres
+spring.datasource.password=12345
 
 #Jpa conf
 spring.jpa.hibernate.ddl-auto=none
@@ -9,7 +9,7 @@ spring.jpa.database-platform=org.hibernate.dialect.PostgreSQL10Dialect
 spring.jpa.show-sql=true
 
 #Server port
-server.port=5000
+server.port=9090
 
 #JWT
 app.jwtSecret= SecretKey
